refactor(streams): merge streams2 prototype into streams
The streams2 module duplicated both the IStream interface hierarchy and the Step
engine. Since the engine was already ported into streams in a more complete form
(Pending with async leaves, recover/doOnError, the suspending driver, batch
chunking) and nothing depended on streams2, the separate module was pure
redundancy.

- Remove the streams2 module and its settings.gradle.kts registration.
- Keep its batching/dedup/stack-safety tests, rewritten against the real
  BulkRequestStreamExecutor / enqueue API (BulkRequestBatchingTest).
- Add streams/README.md with the design overview (previously in streams2/README).
- Update streams-redesign.md to describe the prototype-then-merge history.
