+++ settings.gradle.kts
@@ -41,6 +41,5 @@ include("mps-multiplatform-lib")
 include("mps-repository-concepts")
 include("mps-sync-plugin3")
 include("streams")
-include("streams2")
 include("ts-model-api")
 include("vue-model-api")
