--- streams2/build.gradle.kts
@@ -1,18 +1,0 @@
-plugins {
-    `maven-publish`
-    `modelix-kotlin-multiplatform`
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-            }
-        }
-        commonTest {
-            dependencies {
-                implementation(kotlin("test"))
-            }
-        }
-    }
-}
